% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_pos_to_snp.R
\name{chr_pos_to_snp}
\alias{chr_pos_to_snp}
\title{根据chr 和pos 补充snp}
\usage{
chr_pos_to_snp(data, data_ref = SNPlocs.Hsapiens.dbSNP155.GRCh37)
}
\arguments{
\item{data}{输入data,包含CHR和BP列，需要提前修改列名}

\item{package}{使用的包SNPlocs.Hsapiens.dbSNP155.GRCh37/SNPlocs.Hsapiens.dbSNP155.GRCh38}
}
\value{
data contain SNP
}
\description{
根据chr 和pos 补充snp
}
