% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_pos_to_snp.R
\name{chr_pos_to_snp_batch}
\alias{chr_pos_to_snp_batch}
\title{根据chr 和pos 补充snp}
\usage{
chr_pos_to_snp_batch(
  data,
  data_ref = SNPlocs.Hsapiens.dbSNP155.GRCh37,
  batch = 10^6,
  out_name = "out"
)
}
\arguments{
\item{data}{输入data}

\item{package}{使用的包SNPlocs.Hsapiens.dbSNP155.GRCh37/SNPlocs.Hsapiens.dbSNP155.GRCh38}
}
\value{
data contain snp
}
\description{
根据chr 和pos 补充snp
}
\examples{
#library(SNPlocs.Hsapiens.dbSNP155.GRCh37)
#library(SNPlocs.Hsapiens.dbSNP155.GRCh38)
#library(GenomicRanges)
#data=data.table::fread("AA_invEic_15_oxoEDE_a_asgpcepc.txt")
#colnames(data) <- c(
#  "chromosome", 
#  "base_pair_location",
#  "effect_allele",
#  "other_allele",
#  "beta",
#  "sebeta",
#  "effect_allele_frequency",
#  "p_value",
#  "rsid"
#)
#data.table::fwrite(data,"AA_invEic_15_oxoEDE_a_asgpcepc.csv")
#library(MungeSumstats)
#
#outcome=data
#
#colnames(outcome)
#outcome=outcome[,c(1,2,3,4,8,5,6)]
#head(outcome)
#colnames(outcome)=c( "CHR" , "BP" ,  "A1" ,  "A2"  , "P"  ,  "BETA", "SE"  )#
##' #a=chr_pos_to_snp_batch(data=outcome,out_name = "AA_invEic_15_oxoEDE_a_asgpcepc_snp.csv",batch = 3e4,data_ref = SNPlocs.Hsapiens.dbSNP155.GRCh37)
##' #转换率：0.3372293897441
#b=chr_pos_to_snp_batch(data=outcome,out_name = "AA_invEic_15_oxoEDE_a_asgpcepc_snp.csv",batch = 3e4,data_ref = SNPlocs.Hsapiens.dbSNP155.GRCh38)
##' #转换率：0.96612293897441
#d=b[b$SNP=="",-8]
#rm("outcome")
#rm("data")
#gc()
#data.table::fwrite(d,"d.csv")
#d=data.table::fread("d.csv")
#d=chr_pos_to_snp_batch(data=d,out_name = "AA_invEic_15_oxoEDE_a_asgpcepc_snp_.csv",batch = 1e4,data_ref = SNPlocs.Hsapiens.dbSNP155.GRCh37)
#process_chunked_sumstats(d,dbSNP = "155",ref_genome = "GRCh38")
}
