% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factpc.R
\name{factpc}
\alias{factpc}
\title{simple function}
\usage{
factpc(X, m = 2, rotation = "none", scores = "regression")
}
\arguments{
\item{x}{x is a matrix}
}
\value{
a dataframe
}
\description{
today,I create my first function,a very usrful function.
}
\details{
This is some descriptio of this function.

you can use this function to caculate x+1,then return the value of x+1.
}
