% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_NA.R
\name{find_NA}
\alias{find_NA}
\title{simple function}
\usage{
find_NA(data = data, m = m)
}
\arguments{
\item{data}{is dataframe,m is variable which}
}
\description{
cha zhao que shi zhi wei zhi ji que shi ge shu
}
\details{
This is some descriptio of this function.
}
