% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_csv_TwoSampleMR.R
\name{get_eaf_from_1000G}
\alias{get_eaf_from_1000G}
\title{Title}
\usage{
get_eaf_from_1000G(dat, path, type = "exposure")
}
\arguments{
\item{dat}{}

\item{path}{}

\item{type}{}
}
\description{
Title
}
\examples{
#maffile_path="E:\\wd\\MR\\ld_clump_local\\bfile\\"
#exp1=get_eaf_from_1000G(exp,path = "E:\\wd\\MR\\ld_clump_local\\bfile\\EUR_freq.frq")
}
