% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.text.R
\name{plot.text}
\alias{plot.text}
\title{simple function}
\usage{
\method{plot}{text}(X, h = 0, v = 0)
}
\arguments{
\item{x}{x is a matrix}
}
\value{
a dataframe
}
\description{
today,I create my first function,a very usrful function.
}
\details{
This is some descriptio of this function.

you can use this function to caculate x+1,then return the value of x+1.
}
