% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chunked_sumstats .R
\name{process_chunked_sumstats}
\alias{process_chunked_sumstats}
\title{根据chr 和pos 补充snp}
\usage{
process_chunked_sumstats(
  outcome,
  chunk_size = 10000,
  dbSNP = "144",
  ref_genome = "GRCh37",
  save_path = "out.csv.gz"
)
}
\arguments{
\item{outcome}{输入outcome文件路径}

\item{chunk_size}{分块读取行数（默认10000）}

\item{dbSNP}{dbSNP版本（默认"144"，当前仅占位）}

\item{ref_genome}{参考基因组版本（默认"GRCh37"，当前仅占位）}

\item{save_path}{输出CSV.gz文件路径（默认"out.csv.gz"）}
}
\value{
无返回值，结果保存至指定文件
}
\description{
处理分块读取的VCF.gz文件，提取关键统计信息并保存为TwoSampleMR兼容的CSV.gz格式
}
\examples{
# 执行分块处理
# outcome=fread("outcome_chr_pos.csv")
# colnames(outcome)
# outcome=outcome[,c(16,17,2,3,10,8,9)]
# head(outcome)
# process_chunked_sumstats(
#   outcome = outcome,chunk_size = 10^6,
#   dbSNP = "144",
#   ref_genome = "GRCh37",
#   save_path = "out.csv.gz"  # 输出到results/sumstats_part1.csv.gz等
# )
}
