% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_TwoAampleMR.R
\name{SNP2eaf}
\alias{SNP2eaf}
\title{SNP2eaf}
\usage{
SNP2eaf(
  dat,
  population = c("gnomADg:nfe", "gnomADg:ami", "gnomADg:amr", "gnomADg:asj",
    "gnomADg:afr", "gnomADg:mid", "gnomADg:sas", "gnomADg:ALL", "gnomADg:remaining",
    "gnomADg:eas", "gnomADg:fin", "ALFA:SAMN10492703", "ALFA:SAMN10492705",
    "ALFA:SAMN10492695", "ALFA:SAMN10492696", "ALFA:SAMN10492702", "ALFA:SAMN10492700",
    "ALFA:SAMN11605645", "ALFA:SAMN10492698", "ALFA:SAMN10492699", "GEM-J", "TOPMed",
    "1000GENOMES:phase_3:ACB", "1000GENOMES:phase_3:AFR", "1000GENOMES:phase_3:ALL",
    "1000GENOMES:phase_3:AMR", "1000GENOMES:phase_3:ASW", 
    
    "1000GENOMES:phase_3:BEB", "1000GENOMES:phase_3:CDX", "1000GENOMES:phase_3:CEU",
    "1000GENOMES:phase_3:CHB", "1000GENOMES:phase_3:CHS", "1000GENOMES:phase_3:CLM",
    "1000GENOMES:phase_3:EAS", "1000GENOMES:phase_3:ESN", "1000GENOMES:phase_3:EUR",
    "1000GENOMES:phase_3:FIN", "1000GENOMES:phase_3:GBR", "1000GENOMES:phase_3:GIH",
    "1000GENOMES:phase_3:GWD", "1000GENOMES:phase_3:IBS", "1000GENOMES:phase_3:ITU",
    "1000GENOMES:phase_3:JPT", "1000GENOMES:phase_3:KHV", "1000GENOMES:phase_3:LWK",
    "1000GENOMES:phase_3:MSL", 
     "1000GENOMES:phase_3:MXL",
    "1000GENOMES:phase_3:PEL", "1000GENOMES:phase_3:PJL", "1000GENOMES:phase_3:PUR",
    "1000GENOMES:phase_3:SAS", "1000GENOMES:phase_3:STU", "1000GENOMES:phase_3:TSI",
    "1000GENOMES:phase_3:YRI", "GGVP:GWF", "GGVP:GWD", "GGVP:GWW", "GGVP:GWJ",
    "GGVP:ALL")[36],
  build = "37"
)
}
\arguments{
\item{dat}{data of TwoSamplMR}

\item{population}{地区}

\item{build}{37 or 38}
}
\description{
SNP2eaf
}
