% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_TwoAampleMR.R
\name{snp_add_exposure_eaf}
\alias{snp_add_exposure_eaf}
\title{snp_add_exposure_eaf ,根据snp补全全eaf值}
\usage{
snp_add_exposure_eaf(dat, build = "37", pop = "EUR")
}
\arguments{
\item{dat}{the data of TwoSampleMR}

\item{build}{37 or 38}

\item{pop}{地区}
}
\description{
snp_add_exposure_eaf ,根据snp补全全eaf值
}
