% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_data.table_to_TwosampleMR.R
\name{vcf_data.table_to_TwosampleMR}
\alias{vcf_data.table_to_TwosampleMR}
\title{将ieu vcf.gz文件转换成TwoSampleMR格式的csv文件}
\usage{
vcf_data.table_to_TwosampleMR(
  vcf =
    "E:/BaiduNetdiskDownload/孟德尔共定位分析/gwas-gwas本地共定位（可以获得vcf文件的gwas数据)/ieu-a-7.vcf.gz",
  type = "exposure"
)
}
\arguments{
\item{vcf}{the location of vcf file}

\item{type}{outcome or exposure}
}
\description{
将ieu vcf.gz文件转换成TwoSampleMR格式的csv文件
}
\examples{
##vcf_data.table_to_TwosampleMR(vcf=system.file("data/eqtl-a-ENSG00000112715.vcf.gz",package = "MYMR"),
##type="exposure")
}
