% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_csv_TwoSampleMR.R
\name{vcf_to_csv_TwosampleMR}
\alias{vcf_to_csv_TwosampleMR}
\title{将ieu vcf.gz文件转换成TwoSampleMR格式的csv文件，方便后续合并}
\usage{
vcf_to_csv_TwosampleMR(vcf, batch_size = 1e+05, type = "exposure")
}
\arguments{
\item{vcf}{}

\item{batch_size}{}

\item{type}{}

\item{arData}{}
}
\description{
将ieu vcf.gz文件转换成TwoSampleMR格式的csv文件，方便后续合并
}
\examples{
#vcf_to_csv_TwosampleMR(vcf=system.file("data/eqtl-a-ENSG00000112715.vcf.gz",package = "MYMR"),
#type="exposure")
}
