% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_TwoAampleMR.R
\name{vcf_to_TwosampleMR}
\alias{vcf_to_TwosampleMR}
\title{vcf_to_TwosampleMR}
\usage{
vcf_to_TwosampleMR(
  vcf = "E:\\\\wd\\\\MR\\\\data\\\\ebi-a-GCST001475.vcf.gz",
  type = "exposure"
)
}
\arguments{
\item{vcf}{vcf.gz文件路径}

\item{type}{exposure or outcome}
}
\description{
vcf_to_TwosampleMR
}
\examples{
#vcf_to_TwosampleMR(vcf=system.file("data/eqtl-a-ENSG00000112715.vcf.gz",package = "MYMR"),
#type="exposure")
}
