% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_.R
\name{weight_}
\alias{weight_}
\title{simple function}
\usage{
weight_(path = path, k = k, word = word)
}
\arguments{
\item{#说明，path为.gal的路径, k表示我们矩阵的维数，word表示我们判断相邻的值，比如从（1, 2, 3, 4），为了求布尔值，相加就得到我们需要的空间权重矩阵}{}
}
\value{
a dataframe
}
\description{
构建空间权重矩阵
}
\details{
This is some descriptio of this function.

#说明，path为.gal的路径,k表示我们矩阵的维数，word表示我们判断相邻的值，比如从（1,2,3,4），为了求布尔值，相加就得到我们需要的空间权重矩阵
}
\examples{
word=read.csv("D:\\\\迅雷下载\\\\中华人民共和国 (1)\\\\云南.csv")[,1]
r=weight_(path="D:\\\\迅雷下载\\\\中华人民共和国 (1)\\\\云南省.gal",k=31,word=word)
}
